requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pydantic>=1.9.0,<3.0.0
PyYAML>=6.0
